CLIM-1447: MOVING-THINGS-AROUND UNDO
