+++ apps/src/lib/s2d.ts
@@ -1,3 +1,5 @@
+import { sprintf } from '@wordpress/i18n';
+import { S2D_FORECAST_CONVENTIONAL_NB_PERIODS } from '@/lib/constants';
 import {
 	ClimateVariableInterface,
 	ForecastType,
@@ -7,14 +9,243 @@ import {
 	S2DFrequencyTypes,
 } from '@/types/climate-variable-interface';
 import { formatUTCDate, utc } from '@/lib/utils';
+import { formatIntlDate } from '@/lib/format';
 import { __ } from '@/context/locale-provider';
 
 import { assertIsS2DFrequencyType } from '@/types/assertions';
-import type {
-	ExtractS2DDownloadStepFilenameComponent,
-	LocationS2DData,
-	SkillLevelData,
-} from './types';
+
+export type PeriodRange = [Date, Date];
+
+/**
+ * Location-specific S2D forecast data from the API.
+ *
+ * Cutoffs are raw values in the variable's unit (°C, mm/day).
+ * Probabilities are 0–100 percentages.
+ */
+export interface LocationS2DData {
+	/** 20th percentile cutoff, e.g. `5.0` → "< 5.0 °C" */
+	cutoff_unusually_low_p20: number;
+	/** 33rd percentile cutoff, e.g. `1.8` → "Below 1.8 °C" */
+	cutoff_below_normal_p33: number;
+	historical_median_p50: number;
+	/** 66th percentile cutoff, e.g. `5.3` → "Above 5.3 °C" */
+	cutoff_above_normal_p66: number;
+	/** 80th percentile cutoff, e.g. `6.9` → "> 6.9 °C" */
+	cutoff_unusually_high_p80: number;
+	/** 0–100, e.g. `4` → "4% probability of being unusually low" */
+	prob_unusually_low: number;
+	/** 0–100, e.g. `37` → "37% probability of being below normal" */
+	prob_below_normal: number;
+	/** 0–100, e.g. `38` → "38% probability of being near normal" */
+	prob_near_normal: number;
+	/** 0–100, e.g. `26` → "26% probability of being above normal" */
+	prob_above_normal: number;
+	/** 0–100, e.g. `5` → "5% probability of being unusually high" */
+	prob_unusually_high: number;
+	skill_level: number;
+	skill_CRPSS: number;
+}
+
+type PeriodJumpUnit = 'month' | 'year';
+
+/**
+ * Resolve the span of one period for a frequency.
+ *
+ * The returned jump is used by `getPeriodEnd` to calculate the period's inclusive
+ * end boundary. It is distinct from `nbPeriods`, which controls how many `[start, end]`
+ * tuples `getPeriods` returns, and from the step used to start the next tuple.
+ *
+ * @param frequency - The frequency whose period span should be resolved.
+ * @param noThrow - If true, the function will return a default value instead of throwing an error for unsupported frequencies.
+ *
+ * @returns A tuple containing the jump size and unit, such as `[3, 'month']` for seasonal frequency.
+ * @throws An error if the frequency type is not supported.
+ */
+export const resolveFrequencyPeriodJump = (
+	frequency: S2DFrequencyType,
+	noThrow: boolean = true,
+): [number, PeriodJumpUnit] => {
+	let periodJumpSize = 1;
+	let periodJumpSizeUnit: PeriodJumpUnit = 'month';
+	if (frequency === S2DFrequencyTypes.SEASONAL) {
+		// Because a season is 3 months long.
+		periodJumpSize = 3;
+	} else if (isFrequencyTypeDecadal(frequency)) {
+		periodJumpSizeUnit = 'year';
+		periodJumpSize = 5; // There might be some logic to adjust this here
+	} else if (frequency === S2DFrequencyTypes.MONTHLY) {
+		periodJumpSize = 1;
+	} else {
+		if (noThrow) {
+			return [periodJumpSize, periodJumpSizeUnit];
+		}
+		throw new Error(
+			sprintf(
+				__('Unsupported frequency type: %s'),
+				frequency
+			)
+		);
+	}
+
+	return [
+		periodJumpSize,
+		periodJumpSizeUnit,
+	];
+};
+
+
+/**
+ * Calculate the inclusive end date for a period starting at `periodStart`.
+ *
+ * @param periodStart - The first day of the period.
+ * @param frequency - The frequency that determines the period length.
+ * @returns The last day of the period.
+ */
+const getPeriodEnd = (
+	periodStart: Date,
+	frequency: S2DFrequencyType,
+): Date => {
+	const [
+		periodJumpSize,
+		periodJumpSizeUnit,
+	] = resolveFrequencyPeriodJump(frequency, false);
+	const periodEnd = new Date(periodStart);
+	if (periodJumpSizeUnit === 'year') {
+		periodEnd.setUTCFullYear(periodStart.getUTCFullYear() + periodJumpSize);
+	} else {
+		periodEnd.setUTCMonth(periodStart.getUTCMonth() + periodJumpSize);
+	}
+	periodEnd.setUTCDate(0);
+	return periodEnd;
+};
+
+
+/**
+ * Return the time periods for a release date and specific frequency.
+ *
+ * The number and length of time periods depend on the frequency:
+ * - Seasonal frequency: 10 x 3-month periods, at 1-month interval
+ * - Monthly frequency: 3 x 1-month periods, at 1-month interval
+ * - Decadal frequency: 2 x 5-year periods, at 5-year interval
+ *
+ * The first period starts on the same month as the release date. A period start
+ * is always the first day of the month, and a period end is always the last
+ * day of the month.
+ *
+ * All dates are in UTC time.
+ *
+ * @see {@link S2D_FORECAST_CONVENTIONAL_NB_PERIODS} for the number of periods per frequency.
+ *
+ * @example
+ * ```typescript
+ * const releaseDate = new Date('2025-10-15');
+ * const periods = getPeriods(releaseDate, S2DFrequencyTypes.MONTHLY);
+ * // Returned periods are (as array of Date instances): [
+ * //   [2025-10-01, 2025-10-31]
+ * //   [2025-11-01, 2025-11-30]
+ * //   [2025-12-01, 2025-12-31]
+ * // ]
+ * ```
+ *
+ * @param releaseDate - The release date of the data.
+ * @param frequency - The frequency for which to get the periods.
+ * @returns An array of [start, end] date instances for each period.
+ */
+export const getPeriods = (
+	releaseDate: Date,
+	frequency: S2DFrequencyType,
+): PeriodRange[] => {
+	// nbPeriods: In other words; how long the list will be
+	const nbPeriods = S2D_FORECAST_CONVENTIONAL_NB_PERIODS[frequency];
+	const periods: PeriodRange[] = [];
+	const lastPeriod = new Date(
+		// Set to the first day of the month
+		Date.UTC(releaseDate.getUTCFullYear(), releaseDate.getUTCMonth(), 1)
+	);
+
+	try {
+		const [
+			periodJumpSize,
+			periodJumpSizeUnit,
+		] = resolveFrequencyPeriodJump(frequency);
+
+		for (let i = 0; i < nbPeriods; i++) {
+			const periodStart = new Date(lastPeriod);
+			const periodEnd = getPeriodEnd(periodStart, frequency);
+			periods.push([periodStart, periodEnd]);
+			if (periodJumpSizeUnit === 'month') {
+				/**
+				 * Monthly and seasonal periods overlap, so advance the next period by one month.
+				 * This creates a moving window while preserving the configured list length.
+				 */
+				lastPeriod.setUTCMonth(lastPeriod.getUTCMonth() + 1 /* "window of months", moving by 1 month at a time */);
+			} else {
+				// In contrast to a "moving window" of months, we do not want years to overlap.
+				lastPeriod.setUTCFullYear(lastPeriod.getUTCFullYear() + periodJumpSize);
+			}
+		}
+	} catch (error) {
+		console.error(
+			`Error calculating periods for frequency "${frequency}":`,
+			error
+		);
+	}
+
+
+	return periods;
+};
+
+/**
+ * Find the index of the period range that matches a string date range.
+ *
+ * A date range is an array of exactly two strings representing a date, in UTC
+ * time. Each string must be of the form 'YYYY-MM-DD'. The date range is
+ * generally created from the `dateRange` URL parameter.
+ *
+ * To find the matching period, only the first date of the date range is used.
+ *
+ * @param dateRange - The date range to search for. An array of exactly two
+ *   strings.
+ * @param availablePeriods - The period ranges to search in.
+ * @returns - The index of the period range that matches the date range, or
+ *   null if not found.
+ */
+export function findPeriodIndexForDateRange(
+	dateRange: [string, string],
+	availablePeriods: PeriodRange[]
+): number | null {
+	const rangeStart = utc(dateRange[0]);
+
+	if (rangeStart === null) {
+		return null;
+	}
+
+	const foundIndex = availablePeriods.findIndex(
+		(period) => rangeStart.toDateString() === period[0].toDateString()
+	);
+
+	return foundIndex === -1 ? null : foundIndex;
+}
+
+/**
+ * Transform a period range (Date instances) to a date range (strings).
+ *
+ * A date range is two strings representing dates in UTC time. The strings are
+ * of the form 'YYYY-MM-DD'.
+ *
+ * @param periodRange - The period range to transform.
+ * @param dateFormat - The format to use for the date strings.
+ * @returns - The date range as an array of two dates in string.
+ */
+export const formatPeriodRange = (
+	periodRange: PeriodRange,
+	dateFormat = 'yyyy-MM-dd',
+): [string, string] => {
+	const rangeStart = formatUTCDate(periodRange[0], dateFormat);
+	const rangeEnd = formatUTCDate(periodRange[1], dateFormat);
+
+	return [rangeStart, rangeEnd];
+};
 
 /**
  * Create and return the GeoServer layer name for the Skill layer.
@@ -217,6 +448,27 @@ export function normalizeForApiFrequencyName(
 	return frequencyNameMap[frequency] ?? frequency;
 }
 
+/**
+ * The extracted S2D filename components from a climate variable that's based on
+ * internal IDs into human relatable strings.
+ */
+export interface ExtractS2DDownloadStepFilenameComponent {
+	/**
+	 * The filename component used to represent the climate variable.
+	 * @see {@link S2D_DOWNLOAD_FILENAME_MAP_VARIABLE_ID}
+	 */
+	variableId: string;
+	/**
+	 * The filename component used to describe the forecast type.
+	 * @see {@link S2D_DOWNLOAD_FILENAME_MAP_FORECAST_TYPE}
+	 */
+	forecastType: string;
+	/**
+	 * The filename component used to describe the frequency.
+	 * @see {@link S2D_DOWNLOAD_FILENAME_MAP_FREQUENCY_TYPE}
+	 */
+	frequencyType: string;
+}
 
 /**
  * Extract and map S2D filename components from a climate variable.
@@ -258,6 +510,11 @@ export const extractS2DDownloadStepFilenameComponents = (
 	};
 };
 
+export interface SkillLevelData {
+	skillLevel: number | null;
+	skillCRPSS: number | null;
+	skillLevelLabel: string;
+}
 
 export const extractSkillLevelData = (
 	locationData: LocationS2DData | null,
@@ -276,3 +533,51 @@ export const extractSkillLevelData = (
 	return output;
 };
 
+/**
+ * Generate a period range label for a given date range and frequency.
+ *
+ * @param dateRangeStart - Start date of the period. Example: 2025-08-01
+ * @param frequency - Frequency type
+ * @param locale - Locale to use for formatting
+ */
+export const generatePeriodRangeLabel = (
+	dateRangeStart: string,
+	frequency: S2DFrequencyType,
+	locale: string
+): string | null => {
+	const periodStart = utc(dateRangeStart);
+
+	const isDecadalFrequencyType = isFrequencyTypeDecadal(frequency);
+
+	if (!periodStart) {
+		return null;
+	}
+
+	const periodStartLabel = isDecadalFrequencyType
+		? formatIntlDate(periodStart, locale, { year: 'numeric' })
+		: formatIntlDate(periodStart, locale, { month: 'long' });
+
+	if (frequency === FrequencyType.MONTHLY) {
+		return periodStartLabel;
+	}
+
+	const periodEnd = getPeriodEnd(periodStart, frequency);
+	const periodEndLabel = isDecadalFrequencyType
+		? formatIntlDate(periodEnd, locale, { year: 'numeric' })
+		: formatIntlDate(periodEnd, locale, { month: 'long' });
+
+	return sprintf(__('%s to %s'), periodStartLabel, periodEndLabel);
+};
+
+/**
+ * The way we represent the percent value and cutoff whether its value is one of the probabilities defined by the cutoffs.
+ * The method of trimming decimal values to determine if the value is part of the cutoff.
+ *
+ * @see {@link LocationS2DData} for the cutoffs and probabilities values.
+ * @see {@link getProbabilitiesBarChartColour} for how the normalized percent value is used to determine the colour of the probability bar.
+ */
+export const normalizeProbabilitiesBarChartPercent = (
+	input: Pick<ProgressBarProps, 'percent'>,
+): number => {
+	return Math.round(input.percent);
+};
